fix: did exchange multiuse invites respond in kind
DID Exchange was responding incorrectly to requests, sending unqualified
DIDs with a did_rotate~attach (which is nonsensical). This fixes it to
only send a did_rotate~attach when the DID associated with the
connection is a qualified DID. Additionally, this change also ensures
that the webhook for the new connection request will only be emitted
after the request has been attached to the connection record. Without
this, a race condition can occur for controllers that try to accept the
request before ACA-Py finishes storing the request.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>